testsuite: Move utility functions to client code
Move utility functions to test cases that use them, if they are used
only once.
