+++ testsuite/python/tests_common.py
@@ -277,60 +277,6 @@ def get_histogram(pos, obs_params, coord_system, **kwargs):
         raise ValueError(f"Unknown coord system '{coord_system}'")
     return np.histogramdd(pos, bins=bins, range=extent, **kwargs)
 
-#
-# Analytical Expressions for interactions
-#
-
-# Harmonic bond
-
-
-def harmonic_potential(scalar_r, k, r_0):
-    return 0.5 * k * (scalar_r - r_0)**2
-
-
-def harmonic_force(scalar_r, k, r_0):
-    return -k * (scalar_r - r_0)
-
-# FENE bond
-
-
-def fene_potential(scalar_r, k, d_r_max, r_0):
-    return -0.5 * k * d_r_max**2 * np.log(1 - ((scalar_r - r_0) / d_r_max)**2)
-
-
-def fene_force(scalar_r, k, d_r_max, r_0):
-    return k * (scalar_r - r_0) * d_r_max**2 / \
-        ((scalar_r - r_0)**2 - d_r_max**2)
-
-
-def fene_force2(bond_vector, k, d_r_max, r_0):
-    r = np.linalg.norm(bond_vector)
-    return k * (r - r_0) / (r * (1 - ((r - r_0) / d_r_max)**2)) * \
-        np.array(bond_vector)
-
-# Coulomb bond
-
-
-def coulomb_potential(scalar_r, k, q1, q2):
-    return k * q1 * q2 / scalar_r
-
-
-def coulomb_force(scalar_r, k, q1, q2):
-    return k * q1 * q2 / scalar_r**2
-
-# QUARTIC bond
-
-
-def quartic_force(k0, k1, r, r_cut, scalar_r):
-    if scalar_r > r_cut:
-        return 0.0
-    return - k0 * (scalar_r - r) - k1 * (scalar_r - r)**3
-
-
-def quartic_potential(k0, k1, r, r_cut, scalar_r):
-    if scalar_r > r_cut:
-        return 0.0
-    return 0.5 * k0 * (scalar_r - r)**2 + 0.25 * k1 * (scalar_r - r)**4
 
 # Generic Lennard-Jones
 
@@ -375,229 +321,6 @@ def lj_force(espressomd, r, epsilon, sigma, cutoff, offset=0.):
     f = lj_generic_force(
         espressomd, r, epsilon, sigma, cutoff, offset=offset, generic=False)
     return f
-
-# Lennard-Jones Cosine
-
-
-def lj_cos_potential(r, epsilon, sigma, cutoff, offset):
-    V = 0.
-    r_min = offset + np.power(2., 1. / 6.) * sigma
-    r_cut = cutoff + offset
-    if r < r_min:
-        V = lj_potential(r, epsilon=epsilon, sigma=sigma,
-                         cutoff=cutoff, offset=offset, shift=0.)
-    elif r < r_cut:
-        alpha = np.pi / \
-            (np.power(r_cut - offset, 2) - np.power(r_min - offset, 2))
-        beta = np.pi - np.power(r_min - offset, 2) * alpha
-        V = 0.5 * epsilon * \
-            (np.cos(alpha * np.power(r - offset, 2) + beta) - 1.)
-    return V
-
-
-def lj_cos_force(espressomd, r, epsilon, sigma, cutoff, offset):
-    f = 0.
-    r_min = offset + np.power(2., 1. / 6.) * sigma
-    r_cut = cutoff + offset
-    if r < r_min:
-        f = lj_force(espressomd, r, epsilon=epsilon, sigma=sigma,
-                     cutoff=cutoff, offset=offset)
-    elif r < r_cut:
-        alpha = np.pi / \
-            (np.power(r_cut - offset, 2) - np.power(r_min - offset, 2))
-        beta = np.pi - np.power(r_min - offset, 2) * alpha
-        f = (r - offset) * alpha * epsilon * \
-            np.sin(alpha * np.power(r - offset, 2) + beta)
-    return f
-
-# Lennard-Jones Cosine^2
-
-
-def lj_cos2_potential(r, epsilon, sigma, offset, width):
-    V = 0.
-    r_min = offset + np.power(2., 1. / 6.) * sigma
-    r_cut = r_min + width
-    if r < r_min:
-        V = lj_potential(r, epsilon=epsilon, sigma=sigma,
-                         offset=offset, cutoff=r_cut, shift=0.)
-    elif r < r_cut:
-        V = -epsilon * np.power(np.cos(np.pi /
-                                       (2. * width) * (r - r_min)), 2)
-    return V
-
-
-def lj_cos2_force(espressomd, r, epsilon, sigma, offset, width):
-    f = 0.
-    r_min = offset + np.power(2., 1. / 6.) * sigma
-    r_cut = r_min + width
-    if r < r_min:
-        f = lj_force(espressomd, r, epsilon=epsilon,
-                     sigma=sigma, cutoff=r_cut, offset=offset)
-    elif r < r_cut:
-        f = - np.pi * epsilon * \
-            np.sin(np.pi * (r - r_min) / width) / (2. * width)
-    return f
-
-# Smooth-Step
-
-
-def smooth_step_potential(r, eps, sig, cutoff, d, n, k0):
-    V = 0.
-    if r < cutoff:
-        V = np.power(d / r, n) + eps / \
-            (1 + np.exp(2 * k0 * (r - sig)))
-    return V
-
-
-def smooth_step_force(r, eps, sig, cutoff, d, n, k0):
-    f = 0.
-    if r < cutoff:
-        f = n * d / r**2 * np.power(d / r, n - 1) + 2 * k0 * eps * np.exp(
-            2 * k0 * (r - sig)) / (1 + np.exp(2 * k0 * (r - sig))**2)
-    return f
-
-# BMHTF
-
-
-def bmhtf_potential(r, a, b, c, d, sig, cutoff):
-    V = 0.
-    if r == cutoff:
-        V = a * np.exp(b * (sig - r)) - c * np.power(
-            r, -6) - d * np.power(r, -8)
-    if r < cutoff:
-        V = a * np.exp(b * (sig - r)) - c * np.power(
-            r, -6) - d * np.power(r, -8)
-        V -= bmhtf_potential(cutoff, a, b, c, d, sig, cutoff)
-    return V
-
-
-def bmhtf_force(r, a, b, c, d, sig, cutoff):
-    f = 0.
-    if r < cutoff:
-        f = a * b * np.exp(b * (sig - r)) - 6 * c * np.power(
-            r, -7) - 8 * d * np.power(r, -9)
-    return f
-
-# Morse
-
-
-def morse_potential(r, eps, alpha, cutoff, rmin=0):
-    V = 0.
-    if r < cutoff:
-        V = eps * (np.exp(-2. * alpha * (r - rmin)) -
-                   2 * np.exp(-alpha * (r - rmin)))
-        V -= eps * (np.exp(-2. * alpha * (cutoff - rmin)) -
-                    2 * np.exp(-alpha * (cutoff - rmin)))
-    return V
-
-
-def morse_force(r, eps, alpha, cutoff, rmin=0):
-    f = 0.
-    if r < cutoff:
-        f = 2. * np.exp((rmin - r) * alpha) * \
-            (np.exp((rmin - r) * alpha) - 1) * alpha * eps
-    return f
-
-# Buckingham
-
-
-def buckingham_potential(r, a, b, c, d, cutoff, discont, shift):
-    V = 0.
-    if r < discont:
-        m = - buckingham_force(
-            discont, a, b, c, d, cutoff, discont, shift)
-        c = buckingham_potential(
-            discont, a, b, c, d, cutoff, discont, shift) - m * discont
-        V = m * r + c
-    if (r >= discont) and (r < cutoff):
-        V = a * np.exp(- b * r) - c * np.power(
-            r, -6) - d * np.power(r, -4) + shift
-    return V
-
-
-def buckingham_force(r, a, b, c, d, cutoff, discont, shift):
-    f = 0.
-    if r < discont:
-        f = buckingham_force(
-            discont, a, b, c, d, cutoff, discont, shift)
-    if (r >= discont) and (r < cutoff):
-        f = a * b * np.exp(- b * r) - 6 * c * np.power(
-            r, -7) - 4 * d * np.power(r, -5)
-    return f
-
-# Soft-sphere
-
-
-def soft_sphere_potential(r, a, n, cutoff, offset=0):
-    V = 0.
-    if r < offset + cutoff:
-        V = a * np.power(r - offset, -n)
-    return V
-
-
-def soft_sphere_force(r, a, n, cutoff, offset=0):
-    f = 0.
-    if (r > offset) and (r < offset + cutoff):
-        f = n * a * np.power(r - offset, -(n + 1))
-    return f
-
-# Hertzian
-
-
-def hertzian_potential(r, eps, sig):
-    V = 0.
-    if r < sig:
-        V = eps * np.power(1 - r / sig, 5. / 2.)
-    return V
-
-
-def hertzian_force(r, eps, sig):
-    f = 0.
-    if r < sig:
-        f = 5. / 2. * eps / sig * np.power(1 - r / sig, 3. / 2.)
-    return f
-
-# Gaussian
-
-
-def gaussian_potential(r, eps, sig, cutoff):
-    V = 0.
-    if r < cutoff:
-        V = eps * np.exp(-np.power(r / sig, 2) / 2)
-    return V
-
-
-def gaussian_force(r, eps, sig, cutoff):
-    f = 0.
-    if r < cutoff:
-        f = eps * r / sig**2 * np.exp(-np.power(r / sig, 2) / 2)
-    return f
-
-
-def gay_berne_potential(r_ij, u_i, u_j, epsilon_0, sigma_0, mu, nu, k_1, k_2):
-    r_normed = r_ij / np.linalg.norm(r_ij)
-    r_u_i = np.dot(r_normed, u_i)
-    r_u_j = np.dot(r_normed, u_j)
-    u_i_u_j = np.dot(u_i, u_j)
-
-    chi = (k_1**2 - 1.) / (k_1**2 + 1.)
-    chi_d = (k_2**(1. / mu) - 1) / (k_2**(1. / mu) + 1)
-
-    sigma = sigma_0 \
-        / np.sqrt(
-            (1 - 0.5 * chi * (
-                (r_u_i + r_u_j)**2 / (1 + chi * u_i_u_j) +
-                (r_u_i - r_u_j)**2 / (1 - chi * u_i_u_j))))
-
-    epsilon = epsilon_0 *\
-        (1 - chi**2 * u_i_u_j**2)**(-nu / 2.) *\
-        (1 - chi_d / 2. * (
-            (r_u_i + r_u_j)**2 / (1 + chi_d * u_i_u_j) +
-            (r_u_i - r_u_j)**2 / (1 - chi_d * u_i_u_j)))**mu
-
-    rr = np.linalg.norm((np.linalg.norm(r_ij) - sigma + sigma_0) / sigma_0)
-
-    return 4. * epsilon * (rr**-12 - rr**-6)
 
 
 def count_fluid_nodes(lbf):
